Don't stop package loading if a single model failed
So far, the loading of the entire package stopped in case a
model could not be loaded or a parser error occurred. This caused
several valid models to be missed.
There is nothing dangerous to continue under this condition. Invalid
CSLPlane_t objects are marked properly and are ignored later.