Resource Manager class
ResourceManager is the central point to get texture and model resources.
In case a resource was already loaded and is still in memory, a
shared pointer is returned.
In case the resource was not yet loaded, async loading will be triggered.
Until the resource loading is completed an empty pointer will be
returned.
ResourceManager itself keeps weak pointers of each loaded resource. As
soon as all shared pointers are deleted, the resource will be freed
automatically.