Fallback to model matching in case a model is unknown
It can happen that a model name is unknown when creating a plane with
model name. The alternatives in this case are:
* Skip creating the plane and don't render anything. Calling code could
handle it and try again.
* Find a good match from the list of loaded planes.

Option 1 can be problematic since the calling code does not know which
planes are loaded and known. Therefore we fall back to libxplanemp's
internal model matching.