Search for more existing LIT texture files before returning the default
This commit unifies the generation of the LIT texture name. Many plane
textures use to have a light texture associated. But not all of them.
They are called as the texture itself but with '_LIT' inserted at the end.
This LIT file path is not defined in the CSL package and several
standards exist, therefore we need to start trying which one exists,
if any.
So far 'LIT' and '_LIT' have been observed in packages. But the list
can be extended.