Filter non existing and already loaded packages when loading new CSL
The cases that CSL_LoadCSL was called with the same inFolderPath
twice was not handled properly. It then tried to reparse everything
and got package name conflicts.
This change filters at an early stage whether the package was already
loaded or if no xsb_aircraft.txt could be found in this folder.